refactor scaffold CLI orchestration
