+++ .sampo/changesets/1057-split-cli-scaffold.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split create scaffold CLI orchestration into focused file, emission, output, and validation helpers.
