+++ packages/wp-typia-project-tools/src/runtime/cli-scaffold.ts
@@ -1,4 +1,3 @@
-import { promises as fsp } from "node:fs";
 import path from "node:path";
 
 import {
@@ -10,31 +9,31 @@ import {
 	isPersistencePolicy,
 	resolvePackageManagerId,
 	resolveTemplateId,
-	scaffoldProject,
 } from "./scaffold.js";
-import { parseAlternateRenderTargets } from "./alternate-render-targets.js";
 import { parseCompoundInnerBlocksPreset } from "./compound-inner-blocks.js";
 import { isCompoundPersistenceEnabled } from "./scaffold-template-variable-groups.js";
 import {
-	formatInstallCommand,
-	formatRunScript,
-} from "./package-managers.js";
-import type {
-	DataStorageMode,
-	PersistencePolicy,
-	ScaffoldProgressEvent,
-} from "./scaffold.js";
-import type { PackageManagerId } from "./package-managers.js";
-import { getPrimaryDevelopmentScript } from "./local-dev-presets.js";
-import { createManagedTempRoot } from "./temp-roots.js";
+	buildScaffoldDryRunPlan,
+	emitScaffoldProject,
+	type ScaffoldEmissionOptions,
+	type ScaffoldInstallDependencies,
+} from "./cli-scaffold-emission.js";
+import { readGeneratedPackageScripts } from "./cli-scaffold-files.js";
 import {
-	getOptionalOnboardingNote,
-	getOptionalOnboardingShortNote,
-	getOptionalOnboardingSteps,
-} from "./scaffold-onboarding.js";
-import { formatNonEmptyTargetDirectoryError } from "./scaffold-bootstrap.js";
-import { pathExists } from "./fs-async.js";
-import { readJsonFile } from "./json-utils.js";
+	getNextSteps,
+	getOptionalOnboarding,
+} from "./cli-scaffold-output.js";
+import {
+	collectProjectDirectoryWarnings,
+	collectTemplateCapabilityWarnings,
+	resolveOptionalBooleanFlag,
+	resolveOptionalSelection,
+	templateUsesPersistenceSettings,
+	validateCreateFlagContract,
+	validateCreateProjectInput,
+} from "./cli-scaffold-validation.js";
+import type { DataStorageMode, PersistencePolicy } from "./scaffold.js";
+import type { PackageManagerId } from "./package-managers.js";
 import {
 	OFFICIAL_WORKSPACE_TEMPLATE_PACKAGE,
 	isBuiltInTemplateId,
@@ -45,36 +44,17 @@ import {
 } from "./external-layer-selection.js";
 import type { TemplateDefinition } from "./template-registry.js";
 import {
-	assertBuiltInTemplateVariantAllowed,
 	resolveLocalCliPathOption,
 	normalizeOptionalCliString,
 } from "./cli-validation.js";
 
-interface GetNextStepsOptions {
-	noInstall: boolean;
-	packageManager: PackageManagerId;
-	projectDir: string;
-	projectInput: string;
-	templateId: string;
-}
-
-interface GetOptionalOnboardingOptions {
-	availableScripts?: string[];
-	packageManager: PackageManagerId;
-	templateId: string;
-	compoundPersistenceEnabled?: boolean;
-}
-
-interface OptionalOnboardingGuidance {
-	note: string;
-	shortNote: string;
-	steps: string[];
-}
-
-export interface ScaffoldDryRunPlan {
-	dependencyInstall: "skipped-by-flag" | "would-install";
-	files: string[];
-}
+export { getNextSteps, getOptionalOnboarding } from "./cli-scaffold-output.js";
+export type {
+	OptionalOnboardingGuidance,
+	ScaffoldNextStepsOptions,
+	ScaffoldOptionalOnboardingOptions,
+} from "./cli-scaffold-output.js";
+export type { ScaffoldDryRunPlan } from "./cli-scaffold-emission.js";
 
 interface RunScaffoldFlowOptions {
 	allowExistingDir?: boolean;
@@ -84,12 +64,12 @@ interface RunScaffoldFlowOptions {
 	dryRun?: boolean;
 	externalLayerId?: string;
 	externalLayerSource?: string;
-	installDependencies?: Parameters<typeof scaffoldProject>[0]["installDependencies"];
+	installDependencies?: ScaffoldInstallDependencies;
 	innerBlocksPreset?: string;
 	isInteractive?: boolean;
 	namespace?: string;
 	noInstall?: boolean;
-	onProgress?: ((event: ScaffoldProgressEvent) => void | Promise<void>) | undefined;
+	onProgress?: ScaffoldEmissionOptions["onProgress"];
 	packageManager?: string;
 	phpPrefix?: string;
 	profile?: string;
@@ -114,474 +94,6 @@ interface RunScaffoldFlowOptions {
 	withTestPreset?: boolean;
 	withWpEnv?: boolean;
 	yes?: boolean;
-}
-
-async function listRelativeProjectFiles(rootDir: string): Promise<string[]> {
-	const relativeFiles: string[] = [];
-
-	async function visit(currentDir: string): Promise<void> {
-		const entries = await fsp.readdir(currentDir, { withFileTypes: true });
-		for (const entry of entries) {
-			const absolutePath = path.join(currentDir, entry.name);
-			if (entry.isDirectory()) {
-				await visit(absolutePath);
-				continue;
-			}
-
-			relativeFiles.push(
-				path
-					.relative(rootDir, absolutePath)
-					.replace(path.sep === "\\" ? /\\/gu : /\//gu, "/"),
-			);
-		}
-	}
-
-	await visit(rootDir);
-	return relativeFiles.sort((left, right) => left.localeCompare(right));
-}
-
-async function assertDryRunTargetDirectoryReady(
-	projectDir: string,
-	allowExistingDir: boolean,
-): Promise<void> {
-	if (!(await pathExists(projectDir)) || allowExistingDir) {
-		return;
-	}
-
-	const entries = await fsp.readdir(projectDir);
-	if (entries.length > 0) {
-		throw new Error(formatNonEmptyTargetDirectoryError(projectDir));
-	}
-}
-
-async function buildScaffoldDryRunPlan({
-	allowExistingDir,
-	alternateRenderTargets,
-	answers,
-	cwd,
-	dataStorageMode,
-	externalLayerId,
-	externalLayerSource,
-	externalLayerSourceLabel,
-	installDependencies,
-	noInstall,
-	onProgress,
-	packageManager,
-	persistencePolicy,
-	profile,
-	projectDir,
-	templateId,
-	variant,
-	withMigrationUi,
-	withTestPreset,
-	withWpEnv,
-}: {
-	allowExistingDir: boolean;
-	alternateRenderTargets?: Parameters<typeof scaffoldProject>[0]["alternateRenderTargets"];
-	answers: Parameters<typeof scaffoldProject>[0]["answers"];
-	cwd: string;
-	dataStorageMode?: Parameters<typeof scaffoldProject>[0]["dataStorageMode"];
-	externalLayerId?: string;
-	externalLayerSource?: string;
-	externalLayerSourceLabel?: string;
-	installDependencies?: Parameters<typeof scaffoldProject>[0]["installDependencies"];
-	noInstall: boolean;
-	onProgress?: RunScaffoldFlowOptions["onProgress"];
-	packageManager: PackageManagerId;
-	persistencePolicy?: Parameters<typeof scaffoldProject>[0]["persistencePolicy"];
-	profile?: Parameters<typeof scaffoldProject>[0]["profile"];
-	projectDir: string;
-	templateId: string;
-	variant?: string;
-	withMigrationUi: boolean;
-	withTestPreset: boolean;
-	withWpEnv: boolean;
-}): Promise<{
-	plan: ScaffoldDryRunPlan;
-	result: Awaited<ReturnType<typeof scaffoldProject>>;
-}> {
-	await assertDryRunTargetDirectoryReady(projectDir, allowExistingDir);
-	const { path: tempRoot, cleanup } = await createManagedTempRoot(
-		"wp-typia-scaffold-plan-",
-	);
-	const previewProjectDir = path.join(tempRoot, "preview-project");
-
-	try {
-		const result = await scaffoldProject({
-			allowExistingDir: false,
-			alternateRenderTargets,
-			answers,
-			cwd,
-			dataStorageMode,
-			externalLayerId,
-			externalLayerSource,
-			externalLayerSourceLabel,
-			installDependencies,
-			noInstall: true,
-			onProgress,
-			packageManager,
-			persistencePolicy,
-			profile,
-			projectDir: previewProjectDir,
-			templateId,
-			variant,
-			withMigrationUi,
-			withTestPreset,
-			withWpEnv,
-		});
-		const files = await listRelativeProjectFiles(previewProjectDir);
-
-		return {
-			plan: {
-				dependencyInstall: noInstall ? "skipped-by-flag" : "would-install",
-				files,
-			},
-			result,
-		};
-	} finally {
-		await cleanup();
-	}
-}
-
-function validateCreateProjectInput(projectInput: string) {
-	const normalizedProjectInput = projectInput.trim();
-	if (normalizedProjectInput.length === 0) {
-		throw new Error(
-			"Project directory is required. Usage: wp-typia create <project-dir> (or wp-typia <project-dir> when <project-dir> is the only positional argument).",
-		);
-	}
-
-	const normalizedProjectPath =
-		path.normalize(normalizedProjectInput).replace(/[\\/]+$/u, "") ||
-		path.normalize(normalizedProjectInput);
-	if (normalizedProjectPath === "." || normalizedProjectPath === "..") {
-		throw new Error(
-			"`wp-typia create` requires a new project directory. Use an explicit child directory instead of `.` or `..`.",
-		);
-	}
-}
-
-function collectProjectDirectoryWarnings(projectDir: string): string[] {
-	const warnings: string[] = [];
-	const projectName = path.basename(projectDir);
-	if (/\s/u.test(projectName)) {
-		warnings.push(
-			`Project directory "${projectName}" contains spaces. The generated next-step commands will be quoted, but a simple kebab-case directory name is usually easier to use with shells and downstream tooling.`,
-		);
-	}
-
-	const shellSensitiveCharacters = Array.from(
-		new Set(projectName.match(/[^A-Za-z0-9._ -]/gu) ?? []),
-	);
-	if (shellSensitiveCharacters.length > 0) {
-		warnings.push(
-			`Project directory "${projectName}" contains shell-sensitive characters (${shellSensitiveCharacters.join(", ")}). Prefer letters, numbers, ".", "_" and "-" when possible.`,
-		);
-	}
-
-	return warnings;
-}
-
-function templateUsesPersistenceSettings(
-	templateId: string,
-	options: {
-		dataStorageMode?: string;
-		persistencePolicy?: string;
-	},
-): boolean {
-	if (templateId === "persistence") {
-		return true;
-	}
-
-	if (templateId !== "compound") {
-		return false;
-	}
-
-	return Boolean(options.dataStorageMode || options.persistencePolicy);
-}
-
-function templateSupportsPersistenceFlags(templateId: string): boolean {
-	return templateId === "persistence" || templateId === "compound";
-}
-
-function templateSupportsCompoundInnerBlocksPreset(templateId: string): boolean {
-	return templateId === "compound";
-}
-
-function createTemplateLabel(templateId: string): string {
-	return templateId === OFFICIAL_WORKSPACE_TEMPLATE_PACKAGE
-		? "`--template workspace`"
-		: `"${templateId}"`;
-}
-
-function collectTemplateCapabilityWarnings(options: {
-	queryPostType?: string;
-	templateId: string;
-	withMigrationUi?: boolean;
-}): string[] {
-	const warnings: string[] = [];
-	const trimmedQueryPostType = options.queryPostType?.trim();
-
-	if (
-		trimmedQueryPostType &&
-		options.templateId !== "query-loop" &&
-		(isBuiltInTemplateId(options.templateId) ||
-			options.templateId === OFFICIAL_WORKSPACE_TEMPLATE_PACKAGE)
-	) {
-		warnings.push(
-			`\`--query-post-type\` only applies to \`wp-typia create --template query-loop\`, which scaffolds a create-time \`core/query\` variation instead of a standalone block. ${createTemplateLabel(options.templateId)} will ignore "${trimmedQueryPostType}".`,
-		);
-	}
-
-	if (
-		options.withMigrationUi === true &&
-		!isBuiltInTemplateId(options.templateId) &&
-		options.templateId !== OFFICIAL_WORKSPACE_TEMPLATE_PACKAGE
-	) {
-		warnings.push(
-			`\`--with-migration-ui\` was ignored for ${createTemplateLabel(options.templateId)}. Migration UI currently scaffolds built-in templates and the official \`--template workspace\` flow; external templates still need to opt into that surface explicitly.`,
-		);
-	}
-
-	return warnings;
-}
-
-function templateSupportsAlternateRenderTargets(options: {
-	alternateRenderTargets?: string;
-	dataStorageMode?: string;
-	persistencePolicy?: string;
-	templateId: string;
-}): boolean {
-	if (!options.alternateRenderTargets) {
-		return false;
-	}
-
-	if (options.templateId === "persistence") {
-		return true;
-	}
-
-	if (options.templateId !== "compound") {
-		return false;
-	}
-
-	return templateUsesPersistenceSettings(options.templateId, {
-		dataStorageMode: options.dataStorageMode,
-		persistencePolicy: options.persistencePolicy,
-	});
-}
-
-function validateCreateFlagContract(options: {
-	alternateRenderTargets?: string;
-	dataStorageMode?: string;
-	innerBlocksPreset?: string;
-	persistencePolicy?: string;
-	templateId: string;
-	variant?: string;
-}) {
-	const {
-		alternateRenderTargets,
-		dataStorageMode,
-		innerBlocksPreset,
-		persistencePolicy,
-		templateId,
-		variant,
-	} = options;
-	if (
-		(dataStorageMode || persistencePolicy) &&
-		!templateSupportsPersistenceFlags(templateId)
-	) {
-		throw new Error(
-			"`--data-storage` and `--persistence-policy` are supported only for `wp-typia create --template persistence` or `--template compound`.",
-		);
-	}
-	if (
-		alternateRenderTargets &&
-		!templateSupportsAlternateRenderTargets({
-			alternateRenderTargets,
-			dataStorageMode,
-			persistencePolicy,
-			templateId,
-		})
-	) {
-		if (templateId === "compound") {
-			throw new Error(
-				"`--alternate-render-targets` on `wp-typia create --template compound` requires the persistence-enabled server render path. Add `--data-storage <post-meta|custom-table>` or `--persistence-policy <authenticated|public>` first.",
-			);
-		}
-		throw new Error(
-			"`--alternate-render-targets` is supported only for `wp-typia create --template persistence` or persistence-enabled `--template compound` scaffolds.",
-		);
-	}
-	parseAlternateRenderTargets(alternateRenderTargets);
-	if (
-		innerBlocksPreset &&
-		!templateSupportsCompoundInnerBlocksPreset(templateId)
-	) {
-		throw new Error(
-			"`--inner-blocks-preset` is supported only for `wp-typia create --template compound`.",
-		);
-	}
-	parseCompoundInnerBlocksPreset(innerBlocksPreset);
-
-	if (isBuiltInTemplateId(templateId)) {
-		assertBuiltInTemplateVariantAllowed({
-			templateId,
-			variant,
-		});
-	}
-}
-
-function parseSelectableValue<T extends string>(
-	label: string,
-	value: string,
-	isValue: (input: string) => input is T,
-	allowedValues: readonly T[],
-): T {
-	if (isValue(value)) {
-		return value;
-	}
-
-	throw new Error(
-		`Unsupported ${label} "${value}". Expected one of: ${allowedValues.join(", ")}`,
-	);
-}
-
-async function resolveOptionalSelection<T extends string>({
-	defaultValue,
-	explicitValue,
-	isInteractive,
-	isValue,
-	label,
-	allowedValues,
-	select,
-	shouldResolve = true,
-	yes,
-}: {
-	defaultValue: T;
-	explicitValue?: string;
-	isInteractive: boolean;
-	isValue: (input: string) => input is T;
-	label: string;
-	allowedValues: readonly T[];
-	select?: () => Promise<T>;
-	shouldResolve?: boolean;
-	yes: boolean;
-}): Promise<T | undefined> {
-	if (!shouldResolve) {
-		return undefined;
-	}
-
-	if (explicitValue) {
-		return parseSelectableValue(label, explicitValue, isValue, allowedValues);
-	}
-
-	if (yes) {
-		return defaultValue;
-	}
-
-	if (isInteractive && select) {
-		return select();
-	}
-
-	return defaultValue;
-}
-
-async function resolveOptionalBooleanFlag({
-	defaultValue = false,
-	disabled = false,
-	explicitValue,
-	isInteractive,
-	select,
-	yes,
-}: {
-	defaultValue?: boolean;
-	disabled?: boolean;
-	explicitValue?: boolean;
-	isInteractive: boolean;
-	select?: () => Promise<boolean>;
-	yes: boolean;
-}): Promise<boolean> {
-	if (disabled) {
-		return defaultValue;
-	}
-
-	if (typeof explicitValue === "boolean") {
-		return explicitValue;
-	}
-
-	if (yes) {
-		return defaultValue;
-	}
-
-	if (isInteractive && select) {
-		return select();
-	}
-
-	return defaultValue;
-}
-
-function quoteShellValue(value: string): string {
-	if (
-		!value.startsWith("-") &&
-		/^[A-Za-z0-9._/@:-]+(?:\/[A-Za-z0-9._@:-]+)*$/.test(value)
-	) {
-		return value;
-	}
-
-	return `'${value.replace(/'/g, `'\"'\"'`)}'`;
-}
-
-/**
- * Build the printed next-step commands for a scaffolded project.
- *
- * @param options Project location and package-manager details used to format
- * next-step commands.
- * @returns Ordered shell commands shown after scaffolding succeeds.
- */
-export function getNextSteps({
-	projectInput,
-	projectDir,
-	packageManager,
-	noInstall,
-	templateId,
-}: GetNextStepsOptions): string[] {
-	const cdTarget = path.isAbsolute(projectInput) ? projectDir : projectInput;
-	const steps = [`cd ${quoteShellValue(cdTarget)}`];
-
-	if (noInstall) {
-		steps.push(formatInstallCommand(packageManager));
-	}
-
-	steps.push(formatRunScript(packageManager, getPrimaryDevelopmentScript(templateId)));
-	return steps;
-}
-
-/**
- * Compute optional onboarding guidance shown after scaffolding completes.
- *
- * @param options Package-manager and template context for optional guidance.
- * @returns Optional onboarding note and step list.
- */
-export function getOptionalOnboarding({
-	availableScripts,
-	packageManager,
-	templateId,
-	compoundPersistenceEnabled = false,
-}: GetOptionalOnboardingOptions): OptionalOnboardingGuidance {
-	return {
-		note: getOptionalOnboardingNote(packageManager, templateId, {
-			availableScripts,
-			compoundPersistenceEnabled,
-		}),
-		shortNote: getOptionalOnboardingShortNote(packageManager, templateId, {
-			availableScripts,
-			compoundPersistenceEnabled,
-		}),
-		steps: getOptionalOnboardingSteps(packageManager, templateId, {
-			availableScripts,
-			compoundPersistenceEnabled,
-		}),
-	};
 }
 
 /**
@@ -742,77 +254,38 @@ export async function runScaffoldFlow({
 			answers.compoundInnerBlocksPreset = resolvedInnerBlocksPreset;
 		}
 
+		const emissionOptions = {
+			allowExistingDir,
+			alternateRenderTargets,
+			answers,
+			cwd,
+			dataStorageMode: resolvedDataStorage,
+			externalLayerId: resolvedExternalLayerSelection.externalLayerId,
+			externalLayerSource:
+				resolvedExternalLayerSelection.externalLayerSource,
+			externalLayerSourceLabel: normalizedExternalLayerSource,
+			installDependencies,
+			noInstall,
+			onProgress,
+			packageManager: resolvedPackageManager,
+			persistencePolicy: resolvedPersistencePolicy,
+			profile: resolvedProfile,
+			projectDir,
+			templateId: resolvedTemplateId,
+			variant,
+			withMigrationUi: resolvedWithMigrationUi,
+			withTestPreset: resolvedWithTestPreset,
+			withWpEnv: resolvedWithWpEnv,
+		} satisfies ScaffoldEmissionOptions;
 		const resolvedResult = dryRun
-			? await buildScaffoldDryRunPlan({
-					allowExistingDir,
-					alternateRenderTargets,
-					answers,
-					cwd,
-					dataStorageMode: resolvedDataStorage,
-					externalLayerId: resolvedExternalLayerSelection.externalLayerId,
-					externalLayerSource:
-						resolvedExternalLayerSelection.externalLayerSource,
-					externalLayerSourceLabel: normalizedExternalLayerSource,
-					installDependencies,
-					noInstall,
-					onProgress,
-					packageManager: resolvedPackageManager,
-					persistencePolicy: resolvedPersistencePolicy,
-					profile: resolvedProfile,
-					projectDir,
-					templateId: resolvedTemplateId,
-					variant,
-					withMigrationUi: resolvedWithMigrationUi,
-					withTestPreset: resolvedWithTestPreset,
-					withWpEnv: resolvedWithWpEnv,
-				})
+			? await buildScaffoldDryRunPlan(emissionOptions)
 			: {
 					plan: undefined,
-					result: await scaffoldProject({
-						alternateRenderTargets,
-						answers,
-						allowExistingDir,
-						cwd,
-						dataStorageMode: resolvedDataStorage,
-						externalLayerId: resolvedExternalLayerSelection.externalLayerId,
-						externalLayerSource:
-							resolvedExternalLayerSelection.externalLayerSource,
-						externalLayerSourceLabel: normalizedExternalLayerSource,
-						installDependencies,
-						noInstall,
-						onProgress,
-						packageManager: resolvedPackageManager,
-						persistencePolicy: resolvedPersistencePolicy,
-						profile: resolvedProfile,
-						projectDir,
-						templateId: resolvedTemplateId,
-						variant,
-						withMigrationUi: resolvedWithMigrationUi,
-						withTestPreset: resolvedWithTestPreset,
-						withWpEnv: resolvedWithWpEnv,
-					}),
+					result: await emitScaffoldProject(emissionOptions),
 				};
-		let availableScripts: string[] | undefined;
-		if (!dryRun) {
-			try {
-				const parsedPackageJson = await readJsonFile<{
-					scripts?: unknown;
-				}>(path.join(projectDir, "package.json"), {
-					context: "generated package manifest",
-				});
-				const scripts =
-					parsedPackageJson.scripts &&
-					typeof parsedPackageJson.scripts === "object" &&
-					!Array.isArray(parsedPackageJson.scripts)
-						? parsedPackageJson.scripts
-						: {};
-				availableScripts = Object.entries(scripts)
-					.filter(([, value]) => typeof value === "string")
-					.map(([scriptName]) => scriptName);
-			} catch {
-				availableScripts = undefined;
-			}
-		}
+		const availableScripts = dryRun
+			? undefined
+			: await readGeneratedPackageScripts(projectDir);
 
 		return {
 			dryRun,
