+++ packages/wp-typia-project-tools/tests/cli-entry.test.ts
@@ -141,6 +141,32 @@ describe("@wp-typia/project-tools scaffold CLI flow", () => {
     throw new Error("Expected callback to throw.");
   }
 
+test("cli scaffold flow keeps orchestration separate from file and emission helpers", () => {
+  const runtimeDir = path.join(import.meta.dir, "..", "src", "runtime");
+  const cliScaffold = fs.readFileSync(
+    path.join(runtimeDir, "cli-scaffold.ts"),
+    "utf8"
+  );
+  const scaffoldFiles = fs.readFileSync(
+    path.join(runtimeDir, "cli-scaffold-files.ts"),
+    "utf8"
+  );
+  const scaffoldEmission = fs.readFileSync(
+    path.join(runtimeDir, "cli-scaffold-emission.ts"),
+    "utf8"
+  );
+
+  expect(cliScaffold).toContain('from "./cli-scaffold-emission.js"');
+  expect(cliScaffold).toContain('from "./cli-scaffold-files.js"');
+  expect(cliScaffold).toContain('from "./cli-scaffold-output.js"');
+  expect(cliScaffold).toContain('from "./cli-scaffold-validation.js"');
+  expect(cliScaffold).not.toContain('from "node:fs"');
+  expect(cliScaffold).not.toContain('from "./temp-roots.js"');
+  expect(cliScaffold).not.toContain("readJsonFile");
+  expect(scaffoldFiles).toContain("readGeneratedPackageScripts");
+  expect(scaffoldEmission).toContain("scaffoldProject");
+});
+
 test("CLI diagnostics do not classify unknown template variants as missing templates", () => {
   const diagnostic = serializeCliDiagnosticError(
     createCliCommandError({
